Refactoring based upon PR comments
